11/15/11 Emilio E. Venegas
For this revison I went ahead and modified the addtower() function in the model.h/model.cpp files. Previously the function only accepted one parameter that was the grid number were the tower was going to be placed. With the new modification, the function now takes in two parameters(int), the first parameter keeps its original property, it sets the grid number for the spot were the tower will be placed. The second int tells the function what type of tower the function is to create. The second integer is called an index tower integer. For now there will be three type of towers users can choose. Later we may add towers. I also have to state that for now when the game is played any three of the towers can be chosen, except that their values or tower type index is the same for now. We need to figure out a way to send a different integer depending on what tower the user clicks on the view. Before this revision there was only one tower available, Josh went ahead and adjusted the code to show three different pictures on the view, and also made it possible to reprint the pictures upon the second click(putting the tower were ever the users clicks on the grid) I want to say that in my last submission I said in the comments that I had created the ArrayTowers objects, but I want to clarify that I did not create that object, Sharol and Paola created that object. What I did do today was edit the ArrayTowers.h file to send the tower objects the array holds. I also edited the setTowerTypes() function in the model. Function that was created by Sharol Chand. I edited the function a little bit, but mostly I went ahead and set the three towers that will be created and placed in the ArrayTowers object, then I called the function that places the towers in the array. Once the towers were in the array, I used the setTowerTypeindex() function, which was inherited by the towers from the GraphicsItem object, to set the index of the towers in the area. The index will be accessed and used to identify the type of tower needed to be copied when the user clicks on the tower selection section of the game. Lastly, in order to learn more about the code I messed around with the view code to see how it worked. I want to make clear that the three images that stand for towers are not are permanent images, they are being used for developmental purposes only.