Emilio E. Venegas
This updated code will now allow the user to click on the different towers and create the correct tower object by pulling the tower object from the array and sending it to the tower copy constructor. This is done with the help of the view. When a person clicks on a tower on the view, it sets the tower type index on the view, which is then used by the addTower() function in the model to create the right tower type. There are more precise comments inside the code.

I also was playing around with some of the gui stuff and I added the correct cost values for each tower to be printed out on the view.